% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Do bootstrap replication for a "kmcure" fitted object}
\usage{
bootstrap(
  fitObjName = "fit",
  R = 50,
  useFitEstAsBootInit = TRUE,
  considerPreviousBoots = TRUE,
  silent = FALSE,
  multiOptim_maxit = NULL,
  multiOptim_reltol = NULL,
  optim_reltol = NULL,
  optim_maxit = NULL
)
}
\arguments{
\item{fitObjName}{is the name of a kmcure fit object}

\item{R}{is the number of Bootstrap repeats to be runs}

\item{useFitEstAsBootInit}{a Boolean value which if set to TRUE it use the Fit Estimated coefficients as Initial values in Bootstrap Estimations}

\item{considerPreviousBoots}{a Boolean value which if set to FALSE it ignore possible previously estimated bootstrap replications}

\item{silent}{a Boolean value which if set to TRUE it prevent from showing output messages.}

\item{multiOptim_maxit}{(NULL means to employ the same setting from the kmcure fit object) a number showing the maximum of allowed multiple optimization. The program uses multiple optimization if the convergence of "optim" does not meet.}

\item{multiOptim_reltol}{(NULL means to employ the same setting from the kmcure fit object) a number showing the relative tolerance in continuing multiple optimization procedure.}

\item{optim_reltol}{(NULL means to employ the same setting from the kmcure fit object) a number showing the relative tolerance in continuing of each optimization.}

\item{optim_maxit}{(NULL means to employ the same setting from the kmcure fit object) a number showing the maximum of allowed iterations in each optimization.}
}
\value{
No direct value is returned instead add a boot section in the inputted kmcure fit object
}
\description{
After estimating model parameters using the "kmcure" function we can use the "bootstrap" function to calculate standard errors of the model parameters which are needed for calculating p-values
}
\examples{
# after fitting a model using the "kmcure" function (see ?kmcure for more information),
# we can use the name of the fitted "kmcure" object to do its bootstrap replications:
bootstrap("fit")
}
