% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmcure.R
\name{kmcure}
\alias{kmcure}
\title{Fits AFT Semiparametric Mixture Cure Model using the KME-KDE method}
\usage{
kmcure(
  formula,
  cureform,
  data,
  multiOptim_maxit = 100,
  multiOptim_reltol = 0.001,
  optim_reltol = 1e-08,
  optim_maxit = 500,
  silent = FALSE
)
}
\arguments{
\item{formula}{is the formula of survival sub-model which define time and event variables using Surv function on the left of a tilde symbol and it define list of survival predictors on the right of the tilde symbol.}

\item{cureform}{is the formula of cure sub-model which define predictors of cure sub-model on the right of a tilde symbol, if omitted program use just an intercept in the cure sub-model}

\item{data}{is the input data frame which its columns' names can be called in the formula and the cureform}

\item{multiOptim_maxit}{is the maximum of allowed multiple optimization. If the convergence of "optim" does not meet, the program does multiple optimization.}

\item{multiOptim_reltol}{is the relative tolerance in continuing multiple optimization procedure}

\item{optim_reltol}{is the relative tolerance in continuing of each optimization}

\item{optim_maxit}{is the maximum of allowed iterations in each optimization}

\item{silent}{a Boolean value which if set to TRUE it prevent from showing output messages}
}
\value{
is a kmcure object
}
\description{
Fits AFT (Accelerated Failure Time) Semiparametric Mixture Cure Model using the KME-KDE (Kaplan-Meier Estimation and Kernel Density Estimator) method.
}
\examples{
data(hfp)
fit = kmcure(formula = Surv(TIME, Event) ~
Gender+Smoking+Diabetes+BP+Anaemia+Age+EF+Sodium+Creatinine+Pletelets+CPK,
cureform = ~Gender+Smoking+Diabetes+BP+Anaemia+Age+EF+Sodium+Creatinine+Pletelets+CPK,
data = hfp)
names(fit)
View(fit$mat_coef[, c(1, ncol(fit$mat_coef) )])
}
