% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsa_function.R
\name{lars.lsa}
\alias{lars.lsa}
\title{lars variant for LSA (modifed version that also reports lambda)}
\usage{
lars.lsa(
  Sigma0,
  b0,
  intercept,
  n,
  type = c("lasso", "lar"),
  eps = .Machine$double.eps,
  max.steps
)
}
\arguments{
\item{Sigma0}{The parameter.}

\item{b0}{The intercept of the regression line.}

\item{intercept}{The bool variable of whether consider the intercept situation}

\item{n}{The number of data point.}

\item{type}{Regression options, choose form "lasso" or "lar".}

\item{eps}{The converge threshold defined by the machine.}

\item{max.steps}{The maximum iteration times to stop.}
}
\value{
object.
}
\description{
Solve penalized Least square approximation (LSA) using the Least Angle Regression (LAR) algorithm.
}
\author{
Reference Wang, H. and Leng, C. (2006) and Efron et al. (2004).
}
