% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsa_function.R
\name{lsa}
\alias{lsa}
\title{Ready to use function to apply the lars.lsa function to a fitted object of lm/glm/coxph.}
\usage{
lsa(obj)
}
\arguments{
\item{obj}{lm/glm/coxph or other object.}
}
\value{
beta.ols: the MLE estimate ; beta.bic: the LSA-BIC estimate ; beta.aic: the LSA-AIC estimate.
}
\description{
Ready to use function to apply the lars.lsa function to a fitted object of lm/glm/coxph/etc. This modified version is based on the version of Oct 19, 2006 of the lsa function in the RobMixReg package. Now it also reports lambda, BIC, and AIC.
}
\author{
Reference Wang, H. and Leng, C. (2006) and Efron et al. (2004).
}
