% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAkmcure.R
\name{LSAkmcure}
\alias{LSAkmcure}
\title{LSA of the kmcure fit}
\usage{
LSAkmcure(coef, covmat, n, type)
}
\arguments{
\item{coef}{The estimated coef of a kmcure fit.}

\item{covmat}{The estimated covariance matrix of a kmcure fit}

\item{n}{The number of events.}

\item{type}{variable selection type, choose form "lasso" or "lar".}
}
\value{
object.
}
\description{
LSA of the kmcure fit that need coef, covmat, n, and type
}
