% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized.R
\name{penalized}
\alias{penalized}
\title{Find non-zero penalized coefficients using lars-lsa method}
\usage{
penalized(time, event, survPreds, curePreds = NULL)
}
\arguments{
\item{time}{a survival time to event variable.}

\item{event}{a survival status variable: 1 for event and 0 for censoring.}

\item{survPreds}{a matrix of survival predictor variable(s).}

\item{curePreds}{an optional matrix of curing predictor variable(s).}
}
\value{
a Boolean vector which shows penalized coefficients that are non-zero
}
\description{
Find non-zero penalized coefficients using lars-lsa method
}
