% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmcure.R
\name{predict.kmcure}
\alias{predict.kmcure}
\title{Predict probability of being an Un-cure observation}
\usage{
\method{predict}{kmcure}(fit, newdata = NULL)
}
\arguments{
\item{fit}{a kmcure fit object}

\item{newdata}{an optional matrix with same columns of the model curePreds}
}
\description{
Predict probability of being an Un-cure observation
}
