% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmcure.R
\name{kmcure}
\alias{kmcure}
\title{Title}
\usage{
kmcure(formula, cureform, data)
}
\arguments{
\item{formula}{is the function input}

\item{cureform}{is the function input}

\item{data}{is the function input}
}
\value{
is a kmcure object
}
\description{
Title
}
\examples{
data(hfp)
kmcure(11,22,33)

}
