% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmcure.R
\name{kmcure}
\alias{kmcure}
\title{Fits AFT Semiparametric Mixture Cure Model using the KME-KDE method}
\usage{
kmcure(
  time,
  event,
  survPreds,
  curePreds = NULL,
  multiOptim_maxit = 10,
  multiOptim_reltol = 0.001,
  optim_reltol = 1e-08,
  optim_maxit = 500,
  scale = FALSE,
  silent = FALSE,
  conditional = FALSE,
  optim_method = "Nelder-Mead",
  optim_init = NULL
)
}
\arguments{
\item{time}{a survival time to event variable.}

\item{event}{a survival status variable: 1 for event and 0 for censoring.}

\item{survPreds}{a matrix of survival predictor variable(s).}

\item{curePreds}{an optional matrix of curing predictor variable(s).}

\item{multiOptim_maxit}{a number showing the maximum of allowed multiple optimization. The program uses multiple optimization if the convergence of "optim" does not meet.}

\item{multiOptim_reltol}{a number showing the relative tolerance in continuing multiple optimization procedure.}

\item{optim_reltol}{a number showing the relative tolerance in continuing of each optimization.}

\item{optim_maxit}{a number showing the maximum of allowed iterations in each optimization.}

\item{scale}{a Boolean value that if set to TRUE the program automatically transform predictors using scale function and then back transform the estimated coefficients to the original scale.}

\item{silent}{a Boolean value which if set to TRUE it prevent from showing output messages.}

\item{conditional}{a Boolean value which, if set to TRUE it uses an iterative procedure that estimate parameters survival and cure sub-models conditionally on the last estimation of the other one.}

\item{optim_method}{a string value showing the method of optimization: "Nelder-Mead" and "SANN" are supported.}

\item{optim_init}{an optional numeric vector of initial values. For example, it could be the estimated coefficients of a previous fit to be used in continuing of optimization.}
}
\value{
is a kmcure object
}
\description{
Fits AFT (Accelerated Failure Time) Semiparametric Mixture Cure Model using the KME-KDE (Kaplan-Meier Estimation and Kernel Density Estimator) method.
}
\examples{

data(hfp)

time = hfp$Time

event = hfp$Event

survPreds = hfp[, c(3:15)]
names(survPreds)

curePreds = hfp[, c(3:15)]
names(curePreds)

fit1 = kmcure (time, event, survPreds, curePreds, multiOptim_maxit = 10)

names(fit1)

if(fit1$exitcode==0){
print(fit1$loglik)
print(fit1$timeD)
print(fit1$coef)
}

}
