% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized.R
\name{penalized}
\alias{penalized}
\title{Calculate Penalized AFT cure model which can be used for predictors selection}
\usage{
penalized(time, event, survPreds, curePreds = NULL, R = 100, silent = FALSE)
}
\arguments{
\item{time}{a survival time to event variable.}

\item{event}{a survival status variable: 1 for event and 0 for censoring.}

\item{survPreds}{a matrix of survival predictor variable(s).}

\item{curePreds}{an optional matrix of curing predictor variable(s).}

\item{R}{the number of Bootstrap iterations to be used in this procedure}

\item{silent}{a Boolean value which if set to TRUE it prevent from showing output messages.}
}
\value{
a list showing estimated penalized standard coefficients and non-zoro coefficients
}
\description{
Calculate Penalized AFT cure model which can be used for predictors selection
}
\examples{

data(hfp)

time = hfp$Time

event = hfp$Event

survPreds = hfp[, c(3:15)]
names(survPreds)

curePreds = hfp[, c(3:15)]
names(curePreds)

# penalFit = penalized(time, event, survPreds, curePreds, R = 100)

# names(penalFit)

# penalFit$isNonZeroCoef$lassoBIC

}
